PMSHRI: Issues resolved #111559 and #111602

+++ src/app/views/pm-shri/config/pm_shri_config.ts
@@ -141,7 +141,7 @@ export const config = {
         "options": {
             "bigNumber": {
                 "title": ['Total PM SHRI schools', 'Total districts', 'No. of primary schools', 'No. of upper primary schools', 'No. of secondary schools', 'No. of higher secondary schools', 'KVS', 'NVS'],
-                "formatter": { locale: 'en-IN', format: "long" }
+                "formatter": { locale: 'en-IN', format: "short" }
             }
         }
     }
